move kyma subcommand to kymacmd package
