+++ cmd/kyma/kymacmd/install/cfg_windows.go
