+++ cmd/kyma/kymacmd/install/cmd_test.go
