+++ cmd/kyma/kymacmd/install/hosts_windows.go
