+++ cmd/kyma/kymacmd/test/delete/cmd.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	oct "github.com/kyma-incubator/octopus/pkg/apis/testing/v1alpha1"
-	"github.com/kyma-project/cli/cmd/kyma/test"
+	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/kyma-project/cli/internal/kube"
 	"github.com/kyma-project/cli/pkg/api/octopus"
