+++ cmd/kyma/kymacmd/test/delete/cmd_test.go
