+++ cmd/kyma/kymacmd/test/delete/opts.go
