+++ cmd/kyma/kymacmd/test/list/cmd.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/kyma-project/cli/cmd/kyma/test"
+	"github.com/kyma-project/cli/cmd/kyma/kymacmd/test"
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/kyma-project/cli/internal/kube"
 	"github.com/pkg/errors"
