+++ cmd/kyma/kymacmd/test/status/cmd_test.go
